+++ lib/src/common/export.dart
@@ -1,3 +1,4 @@
 export 'package:able/src/common/able_config.dart';
 export 'package:able/src/common/able_state.dart';
+export 'package:able/src/common/able_type.dart';
 export 'package:able/src/common/able_utils.dart';